Tries to fix crate overlays behaving oddly with tables and curtains
Basically, this happens because vis_contents don't get cut after they get closed, causing it to stay and remain over tables.
This wouldn't be so bad, if it wasn't for timers being responsible for cutting said vis_contents.

This doesn't affect closets, probably because the math's much less complex. Crate code can also probably be simplified, since we aren't doing anything fancy, but I'm not willing to touch it.

As of now, it automatically ends any animation when you move the crate. Which works for all intents and purposes, because it stops the animation when moving into problem areas. (namely, tables and curtains.)